clean up editor
