+++ src/plugins/dashboard/public/application/components/dashboard_editor.tsx
@@ -13,10 +13,6 @@ import { useSavedDashboardInstance } from '../utils/use/use_saved_dashboard_inst
 import { DashboardServices } from '../../types';
 import { useDashboardAppAndGlobalState } from '../utils/use/use_dashboard_app_state';
 import { useEditorUpdates } from '../utils/use/use_editor_updates';
-import {
-  setBreadcrumbsForExistingDashboard,
-  setBreadcrumbsForNewDashboard,
-} from '../utils/breadcrumbs';
 
 export const DashboardEditor = () => {
   const { id: dashboardIdFromUrl } = useParams<{ id: string }>();
@@ -24,8 +20,6 @@ export const DashboardEditor = () => {
   const { chrome } = services;
   const isChromeVisible = useChromeVisibility(chrome);
   const [eventEmitter] = useState(new EventEmitter());
-  const dashboardDom = document.getElementById('dashboardViewport');
-  const [isEmbeddableRendered, setIsEmbeddableRendered] = useState(false);
 
   const { savedDashboard: savedDashboardInstance, dashboard } = useSavedDashboardInstance(
     services,
@@ -41,51 +35,14 @@ export const DashboardEditor = () => {
     dashboard
   );
 
-  const { currentAppState } = useEditorUpdates(
+  const { isEmbeddableRendered, currentAppState } = useEditorUpdates(
     eventEmitter,
+    services,
     dashboard,
     savedDashboardInstance,
     currentContainer,
     appStateContainer
   );
-
-  useEffect(() => {
-    if (currentAppState && dashboard) {
-      if (savedDashboardInstance?.id) {
-        chrome.setBreadcrumbs(
-          setBreadcrumbsForExistingDashboard(
-            savedDashboardInstance.title,
-            currentAppState.viewMode,
-            dashboard.isDirty
-          )
-        );
-        chrome.docTitle.change(savedDashboardInstance.title);
-      } else {
-        chrome.setBreadcrumbs(
-          setBreadcrumbsForNewDashboard(currentAppState.viewMode, dashboard.isDirty)
-        );
-      }
-    }
-  }, [savedDashboardInstance, chrome, currentAppState, dashboard]);
-
-  useEffect(() => {
-    if (!currentContainer || !dashboardDom) {
-      return;
-    }
-    currentContainer.render(dashboardDom);
-    setIsEmbeddableRendered(true);
-
-    return () => {
-      setIsEmbeddableRendered(false);
-    };
-  }, [currentContainer, dashboardDom]);
-
-  useEffect(() => {
-    // clean up all registered listeners if any is left
-    return () => {
-      eventEmitter.removeAllListeners();
-    };
-  }, [eventEmitter]);
 
   return (
     <div>
